@@ -6,7 +6,7 @@ outside of the Mac App Store, with Homebrew cask integration.
 
 from typing import Any
 
-__version__ = "0.1.0-alpha.1"
+__version__ = "0.7.2"
 
 from versiontracker.config import Config, get_config
 from versiontracker.exceptions import (
@@ -42,8 +42,8 @@ __all__ = [
     "parse_version",
     "compare_versions",
     "is_version_newer",
-    "get_config",  # type: ignore[attr-defined]  # Lazy loaded from .config
-    "VersionTrackerError",  # type: ignore[attr-defined]  # Lazy loaded from .exceptions
+    "get_config",
+    "VersionTrackerError",
 ]
 
 
