@@ -64,7 +64,7 @@ def export_data(
             logging.error(f"Error writing to {filename}: {e}")
             raise ExportError(f"Failed to write to {filename}: {e}") from e
     else:
-        return cast(str, content)
+        return content
 
 
 def _export_to_json(
@@ -95,8 +95,8 @@ def _export_to_json(
                         "status": app[2].name if hasattr(app[2], "name") else str(app[2]),
                     }
                 else:
-                    if isinstance(app, tuple) and len(app) > 0:  # type: ignore
-                        app_name = str(app[0])  # type: ignore
+                    if isinstance(app, tuple) and len(app) > 0:
+                        app_name = str(app[0])
                     else:
                         app_name = str(app) if app else ""
                     app_data = {
@@ -118,8 +118,54 @@ def _export_to_json(
         raise ExportError(f"Failed to export to JSON: {e}") from e
 
 
+def _process_applications_dict(writer: Any, applications: list) -> None:
+    """Process applications dict format for CSV export."""
+    for app in applications:
+        if isinstance(app, tuple):
+            app_name = str(app[0]) if len(app) > 0 else ""
+            app_version = str(app[1]) if len(app) > 1 else ""
+            writer.writerow([app_name, app_version, "Unknown", ""])
+
+
+def _process_tuple_list(writer: Any, data: list) -> None:
+    """Process list of tuples format for CSV export."""
+    if data and isinstance(data[0], tuple):
+        for app in data:
+            app_length = len(app)
+            # Check if we have full app data with version info
+            if app_length >= 3 and isinstance(app[1], dict):
+                writer.writerow(
+                    [
+                        str(app[0]),
+                        app[1].get("installed", "") if isinstance(app[1], dict) else "",
+                        (app[1].get("latest", "Unknown") if isinstance(app[1], dict) else "Unknown"),
+                        app[2].name if hasattr(app[2], "name") else str(app[2]),
+                    ]
+                )
+            # Handle minimal tuple case (app name only or with non-dict version info)
+            elif app_length >= 1:
+                writer.writerow([str(app[0]), "", "Unknown", ""])
+
+
+def _process_dict_format(writer: Any, data: dict) -> None:
+    """Process dictionary format for CSV export."""
+    for name, info in data.items():
+        if isinstance(info, dict):
+            writer.writerow(
+                [
+                    name,
+                    info.get("installed", ""),
+                    info.get("latest", "Unknown"),
+                    info.get("status", ""),
+                ]
+            )
+
+
 def _export_to_csv(
-    data: dict[str, Any] | list[tuple[str, dict[str, str], VersionStatus]],
+    data: dict[str, Any]
+    | list[tuple[str, dict[str, str], VersionStatus]]
+    | list[tuple[str, dict[str, str], str]]
+    | list[dict[str, str]],
 ) -> str:
     """Export data to CSV format.
 
@@ -141,51 +187,13 @@ def _export_to_csv(
         # For test data in tests/test_export.py
         # Handle structure like {"applications": [("Firefox", "100.0"), ...]}
         if isinstance(data, dict) and "applications" in data and isinstance(data["applications"], list):
-            for app in data["applications"]:
-                if isinstance(app, tuple):
-                    app_name = str(app[0]) if len(app) > 0 else ""
-                    app_version = str(app[1]) if len(app) > 1 else ""
-                    writer.writerow(
-                        [
-                            app_name,  # name
-                            app_version,  # installed version
-                            "Unknown",  # latest version
-                            "",  # status
-                        ]
-                    )
-
+            _process_applications_dict(writer, data["applications"])
         # Process standard app data (list of tuples)
         elif isinstance(data, list):
-            if data and isinstance(data[0], tuple):
-                for app in data:
-                    app_length = len(app)
-                    # Check if we have full app data with version info
-                    if app_length >= 3 and isinstance(app[1], dict):
-                        writer.writerow(
-                            [
-                                str(app[0]),
-                                app[1].get("installed", "") if isinstance(app[1], dict) else "",
-                                (app[1].get("latest", "Unknown") if isinstance(app[1], dict) else "Unknown"),
-                                app[2].name if hasattr(app[2], "name") else str(app[2]),
-                            ]
-                        )
-                    # Handle minimal tuple case (app name only or with non-dict version info)
-                    elif app_length >= 1:
-                        writer.writerow([str(app[0]), "", "Unknown", ""])
-                    # Note: app_length == 0 would be an empty tuple, which we skip
-
+            _process_tuple_list(writer, data)
         # For dictionary format with app info
         elif isinstance(data, dict):
-            for name, info in data.items():
-                if isinstance(info, dict):
-                    writer.writerow(
-                        [
-                            name,
-                            info.get("installed", ""),
-                            info.get("latest", "Unknown"),
-                            info.get("status", ""),
-                        ]
-                    )
+            _process_dict_format(writer, data)
 
         return output.getvalue()
     except Exception as e:
@@ -193,7 +201,13 @@ def _export_to_csv(
         raise ExportError(f"Failed to export to CSV: {e}") from e
 
 
-def export_to_json(data, filename=None):
+def export_to_json(
+    data: dict[str, Any]
+    | list[tuple[str, dict[str, str], VersionStatus]]
+    | list[tuple[str, dict[str, str], str]]
+    | list[dict[str, str]],
+    filename: str | None = None,
+) -> str:
     """Export data to JSON format.
 
     Args:
@@ -212,12 +226,18 @@ def export_to_json(data, filename=None):
             return filename
         except Exception as e:
             logging.error(f"Error writing to {filename}: {e}")
-            raise ExportError(f"Failed to write to {filename}: {e}")
+            raise ExportError(f"Failed to write to {filename}: {e}") from e
 
     return content
 
 
-def export_to_csv(data, filename=None):
+def export_to_csv(
+    data: dict[str, Any]
+    | list[tuple[str, dict[str, str], VersionStatus]]
+    | list[tuple[str, dict[str, str], str]]
+    | list[dict[str, str]],
+    filename: str | None = None,
+) -> str:
     """Export data to CSV format.
 
     Args:
@@ -231,7 +251,7 @@ def export_to_csv(data, filename=None):
         content = _export_to_csv(data)
     except Exception as e:
         logging.error(f"Error exporting to CSV: {e}")
-        raise ExportError(f"Failed to export to CSV: {e}")
+        raise ExportError(f"Failed to export to CSV: {e}") from e
 
     if filename:
         try:
@@ -240,6 +260,6 @@ def export_to_csv(data, filename=None):
             return filename
         except Exception as e:
             logging.error(f"Error writing to {filename}: {e}")
-            raise ExportError(f"Failed to write to {filename}: {e}")
+            raise ExportError(f"Failed to write to {filename}: {e}") from e
 
     return content
