@@ -27,6 +27,100 @@ from versiontracker.ui import create_progress_bar
 from versiontracker.utils import get_json_data
 
 
+def _get_apps_data() -> list[tuple[str, str]]:
+    """Get application data from system profiler.
+
+    Returns:
+        List of (app_name, version) tuples
+    """
+    print(create_progress_bar().color("green")("Getting application data..."))
+    apps_data = get_json_data(
+        getattr(
+            get_config(),
+            "system_profiler_cmd",
+            "system_profiler -json SPApplicationsDataType",
+        )
+    )
+    return get_applications(apps_data)
+
+
+def _apply_blocklist_filtering(apps: list[tuple[str, str]], options: Any) -> list[tuple[str, str]]:
+    """Apply blocklist/blacklist filtering to applications.
+
+    Args:
+        apps: List of (app_name, version) tuples
+        options: Command line options
+
+    Returns:
+        Filtered list of (app_name, version) tuples
+    """
+    blocklist_value = getattr(options, "blocklist", None)
+    blacklist_value = getattr(options, "blacklist", None)
+
+    if blocklist_value or blacklist_value:
+        # Create a temporary config with the specified blocklist / legacy blacklist
+        temp_config = Config()
+        # Prefer explicit --blocklist over deprecated --blacklist
+        provided = []
+        if blocklist_value and isinstance(blocklist_value, str):
+            provided.extend(blocklist_value.split(","))
+        if blacklist_value and isinstance(blacklist_value, str):
+            # Merge while preserving order; avoid duplicates
+            for item in blacklist_value.split(","):
+                if item not in provided:
+                    provided.append(item)
+        temp_config.set("blacklist", provided)  # config validator currently keyed on 'blacklist'
+        return [(app, ver) for app, ver in apps if not temp_config.is_blocklisted(app)]
+    else:
+        # Use global config for blocklisting
+        return [(app, ver) for app, ver in apps if not get_config().is_blocklisted(app)]
+
+
+def _apply_homebrew_filtering(apps: list[tuple[str, str]], options: Any) -> list[tuple[str, str]]:
+    """Apply Homebrew filtering to applications.
+
+    Args:
+        apps: List of (app_name, version) tuples
+        options: Command line options
+
+    Returns:
+        Filtered list of (app_name, version) tuples
+    """
+    if hasattr(options, "brew_filter") and options.brew_filter:
+        print(create_progress_bar().color("green")("Getting Homebrew casks for filtering..."))
+        brews = get_homebrew_casks()
+        include_brews = getattr(options, "include_brews", False)
+        if not include_brews:
+            return filter_out_brews(apps, brews)
+        else:
+            print(
+                create_progress_bar().color("yellow")("Showing all applications (including those managed by Homebrew)")
+            )
+    return apps
+
+
+def _display_results_table(filtered_apps: list[tuple[str, str]]) -> None:
+    """Display applications in a formatted table.
+
+    Args:
+        filtered_apps: List of (app_name, version) tuples
+    """
+    table = []
+    for app, version in sorted(filtered_apps, key=lambda x: x[0].lower()):
+        table.append(
+            [
+                create_progress_bar().color("green")(app),
+                create_progress_bar().color("blue")(version),
+            ]
+        )
+
+    if table:
+        print(create_progress_bar().color("green")(f"\nFound {len(table)} applications:\n"))
+        print(tabulate(table, headers=["Application", "Version"], tablefmt="pretty"))
+    else:
+        print(create_progress_bar().color("yellow")("\nNo applications found matching the criteria."))
+
+
 def handle_list_apps(options: Any) -> int:
     """Handle listing applications.
 
@@ -46,77 +140,19 @@ def handle_list_apps(options: Any) -> int:
     try:
         logging.info("Starting VersionTracker list command")
 
-        # Get app data
-        print(create_progress_bar().color("green")("Getting application data..."))
+        # Get application data
+        apps = _get_apps_data()
 
-        # Get data from system_profiler
-        apps_data = get_json_data(
-            getattr(
-                get_config(),
-                "system_profiler_cmd",
-                "system_profiler -json SPApplicationsDataType",
-            )
-        )
-
-        # Get applications
-        apps = get_applications(apps_data)
-
-        # Get additional paths if specified
+        # Get additional paths if specified (currently unused)
         if getattr(options, "additional_dirs", None):
             options.additional_dirs.split(",")
 
         # Apply filtering
-        blocklist_value = getattr(options, "blocklist", None)
-        blacklist_value = getattr(options, "blacklist", None)
-
-        if blocklist_value or blacklist_value:
-            # Create a temporary config with the specified blocklist / legacy blacklist
-            temp_config = Config()
-            # Prefer explicit --blocklist over deprecated --blacklist
-            provided = []
-            if blocklist_value and isinstance(blocklist_value, str):
-                provided.extend(blocklist_value.split(","))
-            if blacklist_value and isinstance(blacklist_value, str):
-                # Merge while preserving order; avoid duplicates
-                for item in blacklist_value.split(","):
-                    if item not in provided:
-                        provided.append(item)
-            temp_config.set("blacklist", provided)  # config validator currently keyed on 'blacklist'
-            filtered_apps = [(app, ver) for app, ver in apps if not temp_config.is_blocklisted(app)]
-        else:
-            # Use global config for blocklisting
-            filtered_apps = [(app, ver) for app, ver in apps if not get_config().is_blocklisted(app)]
-
-        # Get Homebrew casks if needed for filtering
-        if hasattr(options, "brew_filter") and options.brew_filter:
-            print(create_progress_bar().color("green")("Getting Homebrew casks for filtering..."))
-            brews = get_homebrew_casks()
-            include_brews = getattr(options, "include_brews", False)
-            if not include_brews:
-                filtered_apps = filter_out_brews(filtered_apps, brews)
-            else:
-                print(
-                    create_progress_bar().color("yellow")(
-                        "Showing all applications (including those managed by Homebrew)"
-                    )
-                )
-
-        # Prepare table data
-        table = []
-        for app, version in sorted(filtered_apps, key=lambda x: x[0].lower()):
-            table.append(
-                [
-                    create_progress_bar().color("green")(app),
-                    create_progress_bar().color("blue")(version),
-                ]
-            )
+        filtered_apps = _apply_blocklist_filtering(apps, options)
+        filtered_apps = _apply_homebrew_filtering(filtered_apps, options)
 
         # Display results
-        if table:
-            print(create_progress_bar().color("green")(f"\nFound {len(table)} applications:\n"))
-            print(tabulate(table, headers=["Application", "Version"], tablefmt="pretty"))
-        else:
-            print(create_progress_bar().color("yellow")("\nNo applications found matching the criteria."))
+        _display_results_table(filtered_apps)
 
         # Export if requested
         if hasattr(options, "export_format") and options.export_format:
@@ -127,7 +163,6 @@ def handle_list_apps(options: Any) -> int:
             )
             if isinstance(export_result, str):
                 print(export_result)
-            return 0
 
         return 0
     except Exception as e:
