@@ -83,6 +83,78 @@ class PerformanceProfiler:
         ps.print_stats(20)  # Top 20 functions by cumulative time
         return s.getvalue()
 
+    def _track_nested_calls(self, name: str) -> bool:
+        """Track nested function calls and return if this is the outermost call.
+
+        Args:
+            name: Function name to track
+
+        Returns:
+            bool: True if this is the outermost call, False if nested
+        """
+        if name not in self._active_functions:
+            self._active_functions.add(name)
+            return True
+        else:
+            self._nested_calls[name] = self._nested_calls.get(name, 0) + 1
+            return False
+
+    def _get_memory_usage(self) -> float:
+        """Get current memory usage in MB.
+
+        Returns:
+            float: Memory usage in MB, or 0 if psutil not available
+        """
+        if HAS_PSUTIL and psutil_module is not None:
+            process = psutil_module.Process()
+            return float(process.memory_info().rss / 1024 / 1024)  # MB
+        return 0.0
+
+    def _update_timing_stats(self, name: str, elapsed: float, memory_before: float, memory_after: float) -> None:
+        """Update timing statistics for a function call.
+
+        Args:
+            name: Function name
+            elapsed: Elapsed time in seconds
+            memory_before: Memory usage before call in MB
+            memory_after: Memory usage after call in MB
+        """
+        memory_diff = memory_after - memory_before
+
+        if name not in self.function_timings:
+            self.function_timings[name] = FunctionTimingInfo(name=name)
+
+        timing = self.function_timings[name]
+        timing.calls += 1
+        timing.total_time += elapsed
+        timing.min_time = min(timing.min_time, elapsed)
+        timing.max_time = max(timing.max_time, elapsed)
+        timing.avg_time = timing.total_time / timing.calls
+        timing.memory_before = memory_before
+        timing.memory_after = memory_after
+        timing.memory_diff = memory_diff
+
+        logging.debug(
+            "Performance: %s took %.4fs, used %.2fMB memory",
+            name,
+            elapsed,
+            memory_diff,
+        )
+
+    def _cleanup_nested_calls(self, name: str, is_outermost_call: bool) -> None:
+        """Clean up tracking data after function execution.
+
+        Args:
+            name: Function name
+            is_outermost_call: Whether this was the outermost call
+        """
+        if is_outermost_call:
+            self._active_functions.remove(name)
+        else:
+            self._nested_calls[name] = self._nested_calls.get(name, 1) - 1
+            if self._nested_calls[name] == 0:
+                del self._nested_calls[name]
+
     def time_function(self, func_name: str | None = None) -> Callable[[Callable[..., T]], Callable[..., T]]:
         """Create a timing decorator for a function.
 
@@ -105,22 +177,9 @@ class PerformanceProfiler:
                 if not self.enabled:
                     return func(*args, **kwargs)
 
-                # Track nested calls to avoid double-counting
-                is_outermost_call = False
-                if name not in self._active_functions:
-                    is_outermost_call = True
-                    self._active_functions.add(name)
-                else:
-                    self._nested_calls[name] = self._nested_calls.get(name, 0) + 1
+                is_outermost_call = self._track_nested_calls(name)
+                memory_before = self._get_memory_usage() if is_outermost_call else 0
 
-                # Only measure memory for outermost calls to avoid skewed measurements
-                if is_outermost_call and HAS_PSUTIL and psutil_module is not None:
-                    process = psutil_module.Process()
-                    memory_before = process.memory_info().rss / 1024 / 1024  # MB
-                else:
-                    memory_before = 0
-
-                # Time the function
                 start_time = time.time()
                 try:
                     result = func(*args, **kwargs)
@@ -128,47 +187,12 @@ class PerformanceProfiler:
                 finally:
                     end_time = time.time()
 
-                    # Only record stats for outermost calls
                     if is_outermost_call:
                         elapsed = end_time - start_time
+                        memory_after = self._get_memory_usage()
+                        self._update_timing_stats(name, elapsed, memory_before, memory_after)
 
-                        # Get final memory usage
-                        if HAS_PSUTIL and psutil_module is not None:
-                            process = psutil_module.Process()
-                            memory_after = process.memory_info().rss / 1024 / 1024  # MB
-                            memory_diff = memory_after - memory_before
-                        else:
-                            memory_after = 0
-                            memory_diff = 0
-
-                        # Update timing info
-                        if name not in self.function_timings:
-                            self.function_timings[name] = FunctionTimingInfo(name=name)
-
-                        timing = self.function_timings[name]
-                        timing.calls += 1
-                        timing.total_time += elapsed
-                        timing.min_time = min(timing.min_time, elapsed)
-                        timing.max_time = max(timing.max_time, elapsed)
-                        timing.avg_time = timing.total_time / timing.calls
-                        timing.memory_before = memory_before
-                        timing.memory_after = memory_after
-                        timing.memory_diff = memory_diff
-
-                        logging.debug(
-                            "Performance: %s took %.4fs, used %.2fMB memory",
-                            name,
-                            elapsed,
-                            memory_diff,
-                        )
-
-                        # Remove from active functions
-                        self._active_functions.remove(name)
-                    else:
-                        # Decrement nested call counter
-                        self._nested_calls[name] = self._nested_calls.get(name, 1) - 1
-                        if self._nested_calls[name] == 0:
-                            del self._nested_calls[name]
+                    self._cleanup_nested_calls(name, is_outermost_call)
 
             return wrapper
 
